Simulated Dataset for Genome-Wide Interaction Analysis

getting_started is a simulated dataset created to demonstrate the use of the sme() function for genome-wide interaction analyses. It contains results from a simulated analysis involving additive genetic effects and gene-by-gene (GxG) interactions.

Usage

data("getting_started")

Format

A list with results from sme(), including the following components:

summary

A data frame summarizing the analysis results, including p-values for SNP associations (p).

pve

A data frame containing the per SNP variance component estimates normalized to phenotypic variance explained (PVE).

vc

A data frame containing the per SNP variance component estimates.

gxg_snps

A vector containing the indices of the SNPs assigned to have epistatic interactions in the trait simulations.

Source

data-raw/getting_started.R

Details

The dataset was generated as follows:

Genotype Simulation: Genotype data for 5000 individuals and 6,000 SNPs was simulated with synthetic allele counts.

Phenotype Simulation: Phenotypic values were simulated with an additive heritability of 0.3 and a GxG interaction heritability of 0.25. A set of 100 SNPs were selected for additive effects, and two groups of 5 SNPs each were used for GxG interactions.

PLINK-Compatible Files: The simulated data was saved in PLINK-compatible .bed, .fam, and .bim files.

Interaction Analysis: The sme() function was used to perform genome-wide interaction analyses on a subset of SNP indices, including the GxG SNP groups and 100 additional additive SNPs. Memory-efficient computation parameters (e.g., chun_ksize, n_randvecs, and n_blocks) were applied.

Key Parameters

Additive Heritability: 0.3

GxG Heritability: 0.25

Number of Samples: 5000

Number of SNPs: 6,000

Selected Additive SNPs: 100

Selected GxG SNP Groups:

Group 1: 5 SNPs

Group 2: 5 SNPs

See also

Examples

data("getting_started")
head(getting_started$summary)
#> # A tibble: 6 × 9
#>   id     index chromosome position         p    pve     vc     se true_gxg_snp
#>   <chr>  <int>      <int>    <int>     <dbl>  <dbl>  <dbl>  <dbl> <lgl>       
#> 1 rs1498  1498          1     1498 0.000581  0.0447 0.0446 0.0137 TRUE        
#> 2 rs2032  2032          1     2032 0.00722   0.0377 0.0377 0.0154 TRUE        
#> 3 rs2364  2364          1     2364 0.00178   0.0450 0.0450 0.0154 TRUE        
#> 4 rs2867  2867          1     2867 0.000496  0.0519 0.0518 0.0157 TRUE        
#> 5 rs4610  4610          1     4610 0.0000783 0.0581 0.0580 0.0153 TRUE        
#> 6 rs822    822          1      822 0.00522   0.0367 0.0367 0.0143 TRUE